chore: rename components for clarity
